requirements.txt
anthropic>=0.25.0
playwright>=1.43.0
gspread>=6.1.0
google-auth>=2.29.0
